% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_auth.R
\name{ao_auth}
\alias{ao_auth}
\title{Authorization in 'AcademyOcean'}
\usage{
ao_auth(
  client_id = ao_get_client_id(),
  client_secret = ao_get_client_secret(),
  token_path = oa_get_token_path()
)
}
\arguments{
\item{client_id}{ID of your App}

\item{client_secret}{Secret of your app}

\item{token_path}{Your directory for credential storage}
}
\value{
Auth object
}
\description{
Authorization in 'AcademyOcean'
}
\seealso{
See \href{https://academyocean.com/api#authorization}{Documentation}
}
