% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_academies.R
\name{ao_get_academies}
\alias{ao_get_academies}
\title{Get the list of all academies. It returns the list of all academies in account with a description of them.}
\usage{
ao_get_academies()
}
\value{
tibble with academies
}
\description{
Get the list of all academies. It returns the list of all academies in account with a description of them.
}
