% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_certificates.R
\name{ao_get_certificates}
\alias{ao_get_certificates}
\title{This returns certificates received by learners}
\usage{
ao_get_certificates(emails, cl = NULL)
}
\arguments{
\item{emails}{Search for a user by email}

\item{cl}{A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).}
}
\value{
tibble with certificates
}
\description{
This returns certificates received by learners
}
