% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_course_progress.R
\name{ao_get_course_progress.default}
\alias{ao_get_course_progress.default}
\title{Method for ao_get_course_progress}
\usage{
\method{ao_get_course_progress}{default}(emails, cl = NULL)
}
\arguments{
\item{emails}{Search for a users by emails}

\item{cl}{A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).}
}
\description{
Method for ao_get_course_progress
}
