% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_course_progress.R
\name{ao_get_course_progress}
\alias{ao_get_course_progress}
\title{This returns a learner’s progress in all active courses/groups.}
\usage{
ao_get_course_progress(emails, cl = NULL)
}
\arguments{
\item{emails}{Search for a users by emails}

\item{cl}{A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).}
}
\value{
tibble with course progress
}
\description{
This returns a learner’s progress in all active courses/groups.
}
