% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners.R
\name{ao_get_leaners}
\alias{ao_get_leaners}
\title{This returns an array of learners from the Academy, or an empty array, if there is no data.}
\usage{
ao_get_leaners()
}
\value{
tibble with leaners
}
\description{
This returns an array of learners from the Academy, or an empty array, if there is no data.
}
