% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners_amount_from_countries.R
\name{ao_get_leaners_amount_from_countries}
\alias{ao_get_leaners_amount_from_countries}
\title{This returns a breakdown of learners by country.}
\usage{
ao_get_leaners_amount_from_countries(country_iso_code = NULL)
}
\arguments{
\item{country_iso_code}{Country iso code}
}
\value{
tibble with leaners geo
}
\description{
This returns a breakdown of learners by country.
}
\examples{
\dontrun{
leaners_stat <- ao_get_leaners_amount_from_countries(
    country_iso_code = "UA"
)
}
}
