% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners_from_country.R
\name{ao_get_leaners_from_country}
\alias{ao_get_leaners_from_country}
\title{This returns a list of learners from the chosen country}
\usage{
ao_get_leaners_from_country(country_iso_code)
}
\arguments{
\item{country_iso_code}{Country iso code}
}
\value{
tibble with leaners
}
\description{
This returns a list of learners from the chosen country
}
\examples{
\dontrun{
ua_leaners <- ao_get_leaners_from_country("UA")
}
}
