% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners_progress_by_course.R
\name{ao_get_leaners_progress_by_course}
\alias{ao_get_leaners_progress_by_course}
\title{This returns a learner’s progress in specified courses}
\usage{
ao_get_leaners_progress_by_course(course_id, course_slug, cl = NULL)
}
\arguments{
\item{course_id}{Course ID, you can get from course settings, from field “API, Course ID”}

\item{course_slug}{Course slug, you can get from course settings, from field ‘course URL’}

\item{cl}{A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).}
}
\value{
When the request is executed, an array of learner’s information is returned
}
\description{
This returns a learner’s progress in specified courses
}
