% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners_registered_at.R
\name{ao_get_leaners_registered_at}
\alias{ao_get_leaners_registered_at}
\title{This returns an array of learners from the Academy, or an empty array, if there is no data.}
\usage{
ao_get_leaners_registered_at(
  date = Sys.Date() - 1,
  operator = c("=", ">", "<", ">=", "<=", "<>")
)
}
\arguments{
\item{date}{Date string for filter}

\item{operator}{One of the following comparison operators '=', '>', '<', '>=', '<=', '<>'}
}
\value{
tibble with leaners
}
\description{
This returns an array of learners from the Academy, or an empty array, if there is no data.
}
