% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners_With_score.R
\name{ao_get_leaners_with_score}
\alias{ao_get_leaners_with_score}
\title{This returns a list of learners filtered by their score}
\usage{
ao_get_leaners_with_score(
  score = 1,
  operator = c(">=", ">", "<", "=", "<=", "<>")
)
}
\arguments{
\item{score}{Filter score the ratio of completed content to available content, across all courses/groups}

\item{operator}{One of the following comparison operators '=', '>', '<', '>=', '<=', '<>'}
}
\value{
tibble with leaners
}
\description{
This returns a list of learners filtered by their score
}
\examples{
\dontrun{
leaners <- ao_get_leaners_with_score(
    score = 20,
    operator = '>'
)
}
}
