% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_learners_churn_at_content.R
\name{ao_get_learners_churn_at_content}
\alias{ao_get_learners_churn_at_content}
\title{This returns an array of learners who churned at a specific lesson in the course.}
\usage{
ao_get_learners_churn_at_content(course_id, content_id)
}
\arguments{
\item{course_id}{Course id, you can get it from edit course, from field ‘course URL’}

\item{content_id}{content id, you can get it from edit content, from field ‘lesson/quiz URL’}
}
\value{
tibble with leaners
}
\description{
This returns an array of learners who churned at a specific lesson in the course.
}
\examples{
\dontrun{
churn <- ao_get_learners_churn_at_content(
    course_id = 'ppc-spetsialist-free',
    content_id = 'itogovyy-test'
)
}
}
