% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_learners_course_complete.R
\name{ao_get_learners_course_complete}
\alias{ao_get_learners_course_complete}
\title{This returns a list of learners who have completed a given course.}
\usage{
ao_get_learners_course_complete(course_id = "GlnLgyVkeqaXkm09Q7Y4")
}
\arguments{
\item{course_id}{Course id, you can get it from edit course, from field ‘course URL’}
}
\value{
tibble with leaners
}
\description{
This returns a list of learners who have completed a given course.
}
\examples{
\dontrun{
leaners <- ao_get_learners_course_complete(
    'ppc-spetsialist-free'
)
}
}
