% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_learners_quiz_statistic.R
\name{ao_get_learners_quiz_statistic}
\alias{ao_get_learners_quiz_statistic}
\title{Returns statistics about the specified quiz for the specified learner(s).}
\usage{
ao_get_learners_quiz_statistic(email, content_id)
}
\arguments{
\item{email}{Search for a users by emails}

\item{content_id}{ID of the quiz you want to see statistics for}
}
\value{
tibble with quize stat
}
\description{
Returns statistics about the specified quiz for the specified learner(s).
}
\examples{
\dontrun{
quize_stat <- ao_get_learners_quiz_statistic(
    email = 'mike.p@gmail.com',
    content_id = 'R9eYq5oE6wLoY3NAPmwx'
)
}
}
