% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_learners_repeated_logins.R
\name{ao_get_learners_repeated_logins}
\alias{ao_get_learners_repeated_logins}
\title{This returns an array of learners who logged in to the Academy more than one time.}
\usage{
ao_get_learners_repeated_logins()
}
\value{
tibble with leaners
}
\description{
This returns an array of learners who logged in to the Academy more than one time.
}
