% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_passive_leaners.R
\name{ao_get_passive_leaners}
\alias{ao_get_passive_leaners}
\title{This returns a list of learners who registered in an Academy but took no further actions (no lessons opened)}
\usage{
ao_get_passive_leaners()
}
\value{
tibble with leaners
}
\description{
This returns a list of learners who registered in an Academy but took no further actions (no lessons opened)
}
