% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_passive_learners.R
\name{ao_get_passive_learners}
\alias{ao_get_passive_learners}
\title{This returns a list of learners who registered in an Academy but took no further actions (no lessons opened)}
\usage{
ao_get_passive_learners()
}
\value{
tibble with passive leaners
}
\description{
This returns a list of learners who registered in an Academy but took no further actions (no lessons opened)
}
\examples{
\dontrun{
passive <- ao_get_passive_learners()
}
}
