% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_auth.R
\name{print.academyocean_token}
\alias{print.academyocean_token}
\title{Print method for academyocean_token}
\usage{
\method{print}{academyocean_token}(x, ..., show_token = FALSE)
}
\arguments{
\item{x}{Auth object}

\item{...}{Using in default print()}

\item{show_token}{Hide or shoh access token in console}
}
\value{
Using only for output token in console
}
\description{
Print method for academyocean_token
}
