% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_course_progress.R
\name{print.oa_course_progress}
\alias{print.oa_course_progress}
\title{Print oa leaners}
\usage{
\method{print}{oa_course_progress}(x, ...)
}
\arguments{
\item{x}{leaners table}

\item{...}{Using in default print()}
}
\value{
Using only for output token in console
}
\description{
Print oa leaners
}
