% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_get_leaners.R
\name{print.oa_leaners}
\alias{print.oa_leaners}
\title{Print oa leaners}
\usage{
\method{print}{oa_leaners}(x, ...)
}
\arguments{
\item{x}{leaners table}

\item{...}{Using in default print()}
}
\value{
Using only for output token in console
}
\description{
Print oa leaners
}
