% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racademyocean_package.R
\docType{package}
\name{racademyocean-package}
\alias{racademyocean}
\alias{racademyocean-package}
\title{racademyocean: Client for 'AcademyOcean API'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide function for work with 'AcademyOcean API' \url{https://academyocean.com/api}.
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
  \item AcademyOcean [copyright holder]
}

}
