% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racademyocean_package.R
\docType{package}
\name{racademyocean-package}
\alias{racademyocean}
\alias{racademyocean-package}
\title{racademyocean: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
  \item AcademyOcean [copyright holder]
}

}
